delete director settting added

+++ WebApi/Application/DirectorOperations/Commands/Delete/DeleteDirectorCommand.cs
@@ -16,6 +16,9 @@ namespace WebApi.Application.DirectorOperations.Commands.DeleteDirector
             if (_Director is null)
                 throw new InvalidOperationException("Director bulunamadı.");
 
+            var _movie = _context.Movies.FirstOrDefault(x => x.DirectorID == DirectorID);
+            if (_movie is not null)
+                throw new InvalidOperationException("Director silinemez. Mevcut Filmi Var.");
 
             _context.Directors.Remove(_Director);
             _context.SaveChanges();
